# docs: update performance benchmarks with latest results