--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,7 +1,7 @@
 
 [package]
 name = "rtest"
-version = "1.0.1"
+version = "1.1.0"
 edition = "2021"
 
 [lib]