--- a/pyproject.toml
+++ b/pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "maturin"
 
 [project]
 name = "rtest"
-version = "1.0.1"
+version = "1.0.2"
 description = "A fast Python test runner built with Rust"
 readme = "README.md"
 requires-python = ">=3.9"