## Summary
- Replace simplified benchmark results in README with detailed hyperfine output
- Show comprehensive performance data for both full test execution and collection-only modes
- Update performance metrics showing significant improvements: 60.52x faster for full execution, 99.61x faster for collection

## Changes
- Updated README.md performance section with latest benchmark results from `./benchmark.sh --all`
- No functional changes to benchmark.sh script

## Performance Highlights
- Full test execution: 3.990s (pytest) vs 65.9ms (rtest) - **60.52x faster**
- Test collection only: 4.051s (pytest) vs 40.7ms (rtest) - **99.61x faster**